requirements.txt
fastapi
uvicorn[standard]
python-dotenv
pydantic

langchain
langgraph
langchain-openai

tiktoken
